+++ kedro-datasets/tests/api/test_api_dataset.py
@@ -701,3 +701,236 @@ class TestAPIDatasetResponseDataset:
         stored_response = api_dataset.get_last_response()
         assert isinstance(stored_response, requests.Response)
         assert stored_response.json() == response_data
+
+
+class TestAPIDatasetSendIndividually:
+    """Test suite for the send_individually feature."""
+
+    @pytest.fixture
+    def mock_json_echo(self, requests_mock):
+        """Fixture to mock API saving and record the request history."""
+        request_history = []
+
+        def json_callback(request: requests.Request, context: Any) -> dict:
+            """Store request and return it."""
+            data = request.json()
+            request_history.append(data)
+            return data
+
+        for method in SAVE_METHODS:
+            requests_mock.register_uri(
+                method,
+                TEST_URL,
+                json=json_callback,
+            )
+        return request_history
+
+    @pytest.mark.parametrize("method", SAVE_METHODS)
+    def test_send_individually_true_sends_each_item(self, mock_json_echo, method):
+        """
+        When send_individually=True with a list of items,
+        Each item should be sent as a separate request.
+        """
+        items = [{"id": 1, "name": "item1"}, {"id": 2, "name": "item2"}]
+
+        api_dataset = APIDataset(
+            url=TEST_URL,
+            method=method,
+            save_args={"send_individually": True},
+        )
+
+        # Save list of items
+        response = api_dataset.save.__wrapped__(api_dataset, items)
+
+        # Verify response is from the last item
+        assert isinstance(response, requests.Response)
+        assert response.json() == items[-1]
+
+        # Verify each item was sent individually (not as a list)
+        assert len(mock_json_echo) == len(items)
+        assert mock_json_echo[0] == items[0]
+        assert mock_json_echo[1] == items[1]
+
+    @pytest.mark.parametrize("method", SAVE_METHODS)
+    def test_send_individually_false_uses_chunking(self, mock_json_echo, method):
+        """
+        When send_individually=False (default) with a list of items,
+        Items should be sent as chunks according to chunk_size.
+        """
+        items = [
+            {"id": 1, "name": "item1"},
+            {"id": 2, "name": "item2"},
+            {"id": 3, "name": "item3"},
+        ]
+
+        api_dataset = APIDataset(
+            url=TEST_URL,
+            method=method,
+            save_args={"send_individually": False, "chunk_size": 2},
+        )
+
+        # Save list of items
+        api_dataset.save.__wrapped__(api_dataset, items)
+
+        # Verify response is from the last chunk
+        assert mock_json_echo[-1] == [items[2]]
+
+        # Verify items were sent as chunks
+        assert len(mock_json_echo) == len(items) - 1
+        assert mock_json_echo[0] == items[:2]  # First chunk: items 1-2
+        assert mock_json_echo[1] == items[2:]  # Second chunk: item 3
+
+    @pytest.mark.parametrize("method", SAVE_METHODS)
+    def test_send_individually_default_false(self, mock_json_echo, method):
+        """
+        When send_individually is not specified,
+        Default behavior should be chunking (send_individually=False).
+        """
+        items = [{"id": 1}, {"id": 2}]
+
+        api_dataset = APIDataset(
+            url=TEST_URL,
+            method=method,
+            save_args={"chunk_size": 1},
+        )
+
+        api_dataset.save.__wrapped__(api_dataset, items)
+
+        # Verify items were sent as chunks (not individually)
+        assert len(mock_json_echo) == len(items)
+        assert mock_json_echo[0] == [items[0]]  # First chunk as list
+        assert mock_json_echo[1] == [items[1]]  # Second chunk as list
+
+    @pytest.mark.parametrize("method", SAVE_METHODS)
+    def test_send_individually_with_single_item(self, mock_json_echo, method):
+        """
+        When send_individually=True with a single item list,
+        The item should be sent individually.
+        """
+        items = [{"id": 1, "data": "single"}]
+
+        api_dataset = APIDataset(
+            url=TEST_URL,
+            method=method,
+            save_args={"send_individually": True},
+        )
+
+        response = api_dataset.save.__wrapped__(api_dataset, items)
+
+        assert isinstance(response, requests.Response)
+        assert len(mock_json_echo) == 1
+        assert mock_json_echo[0] == items[0]
+
+    @pytest.mark.parametrize("method", SAVE_METHODS)
+    def test_send_individually_with_empty_list(self, mock_json_echo, method):
+        """
+        When send_individually=True with an empty list,
+        A DatasetError should be raised.
+        """
+        items = []
+
+        api_dataset = APIDataset(
+            url=TEST_URL,
+            method=method,
+            save_args={"send_individually": True},
+        )
+
+        # Empty list with send_individually=True should raise DatasetError now
+        with pytest.raises(
+            DatasetError, match="Cannot save an empty list with send_individually=True."
+        ):
+            api_dataset.save.__wrapped__(api_dataset, items)
+
+        # No requests should be made for empty list
+        assert len(mock_json_echo) == 0
+
+    @pytest.mark.parametrize("method", SAVE_METHODS)
+    def test_send_individually_with_dict_data(self, mock_json_echo, method):
+        """
+        When send_individually=True but data is a dict (not a list),
+        The dict should be sent as-is.
+        """
+        data = {"id": 1, "name": "single_item"}
+
+        api_dataset = APIDataset(
+            url=TEST_URL,
+            method=method,
+            save_args={"send_individually": True},
+        )
+
+        api_dataset.save.__wrapped__(api_dataset, data)
+
+        # Dict data should still be sent as a single request
+        assert len(mock_json_echo) == 1
+        assert mock_json_echo[0] == data
+
+    @pytest.mark.parametrize("method", SAVE_METHODS)
+    def test_send_individually_takes_precedence_over_chunk_size(
+        self, mock_json_echo, method
+    ):
+        """
+        When both send_individually=True and chunk_size are specified,
+        send_individually should take precedence.
+        """
+        items = [
+            {"id": 1},
+            {"id": 2},
+            {"id": 3},
+            {"id": 4},
+        ]
+
+        api_dataset = APIDataset(
+            url=TEST_URL,
+            method=method,
+            save_args={"send_individually": True, "chunk_size": 2},
+        )
+
+        api_dataset.save.__wrapped__(api_dataset, items)
+
+        # All items should be sent individually (not chunked by 2)
+        assert len(mock_json_echo) == len(items)
+        for i, req_data in enumerate(mock_json_echo):
+            assert req_data == items[i]
+
+    @pytest.mark.parametrize("method", SAVE_METHODS)
+    def test_send_individually_with_response_dataset(
+        self, requests_mock, tmp_path, method
+    ):
+        """
+        When send_individually=True with response_dataset configured,
+        Only the final response should be stored.
+        """
+        items = [
+            {"id": 1, "action": "create"},
+            {"id": 2, "action": "create"},
+            {"id": 3, "action": "create"},
+        ]
+        json_file = tmp_path / "response.json"
+        response_data = {"status": "all_created", "count": 3}
+
+        def json_callback(request: requests.Request, context: Any) -> dict:
+            return response_data
+
+        requests_mock.register_uri(
+            method,
+            TEST_URL,
+            json=json_callback,
+        )
+
+        api_dataset = APIDataset(
+            url=TEST_URL,
+            method=method,
+            save_args={"send_individually": True},
+            response_dataset={"type": "json.JSONDataset", "filepath": str(json_file)},
+        )
+
+        response = api_dataset.save.__wrapped__(api_dataset, items)
+
+        # Verify response is the last one
+        assert response.json() == response_data
+
+        # Verify the final response was stored
+        assert json_file.exists()
+        with open(json_file) as f:
+            stored_data = json.load(f)
+        assert stored_data == response_data
